Clean up unnecessary protocols and remove associated/low-value tests; minor cleanup
