+++ Sources/StytchCore/DFPClient/CAPTCHA.swift
@@ -2,28 +2,22 @@
 import Foundation
 import RecaptchaEnterprise
 
-internal protocol RecaptchaProvider {
-    mutating func configure(siteKey: String) async
+internal protocol CaptchaProvider {
+    func setCaptchaClient(siteKey: String) async
 
-    func getCaptchaToken() async -> String
+    func executeRecaptcha() async -> String
 
     func isConfigured() -> Bool
 }
 
-private struct RecaptchaProviderImplementation: RecaptchaProvider {
+final class CaptchaClient : CaptchaProvider {
     private var recaptchaClient: RecaptchaClient?
 
-    mutating func configure(siteKey: String) async {
-        do {
-            recaptchaClient = try await Recaptcha.getClient(withSiteKey: siteKey)
-        } catch let error as RecaptchaError {
-            print("RecaptchaClient creation error: \(String(describing: error.errorMessage)).")
-        } catch {
-            print("RecaptchaClient creation error: \(String(describing: error))")
-        }
+    func isConfigured() -> Bool {
+        recaptchaClient != nil
     }
 
-    func getCaptchaToken() async -> String {
+    func executeRecaptcha() async -> String {
         guard let recaptchaClient = recaptchaClient else {
             return ""
         }
@@ -38,28 +32,14 @@ private struct RecaptchaProviderImplementation: RecaptchaProvider {
         }
     }
 
-    func isConfigured() -> Bool {
-        recaptchaClient != nil
-    }
-}
-
-final class CAPTCHA {
-    private var recaptchaProvider: RecaptchaProvider
-
-    init(recaptchaProvider: RecaptchaProvider = RecaptchaProviderImplementation()) {
-        self.recaptchaProvider = recaptchaProvider
-    }
-
-    func isConfigured() -> Bool {
-        recaptchaProvider.isConfigured()
-    }
-
-    func executeRecaptcha() async -> String {
-        await recaptchaProvider.getCaptchaToken()
-    }
-
     func setCaptchaClient(siteKey: String) async {
-        await recaptchaProvider.configure(siteKey: siteKey)
+        do {
+            recaptchaClient = try await Recaptcha.getClient(withSiteKey: siteKey)
+        } catch let error as RecaptchaError {
+            print("RecaptchaClient creation error: \(String(describing: error.errorMessage)).")
+        } catch {
+            print("RecaptchaClient creation error: \(String(describing: error))")
+        }
     }
 }
 #endif
