+++ Tests/StytchCoreTests/NetworkingClient+LiveTestCase.swift
@@ -5,19 +5,19 @@ import XCTest
 private class NetworkRequestHandlerMock: NetworkRequestHandler {
     private(set) var methodCalled: String?
     #if os(iOS)
-    func handleDFPDisabled(session _: URLSession, request _: URLRequest, captcha _: CAPTCHA, requestHandler _: (URLSession, URLRequest) async throws -> (Data, HTTPURLResponse)) async throws -> (Data, HTTPURLResponse) {
+    func handleDFPDisabled(session _: URLSession, request _: URLRequest, captcha _: CaptchaProvider, requestHandler _: (URLSession, URLRequest) async throws -> (Data, HTTPURLResponse)) async throws -> (Data, HTTPURLResponse) {
         methodCalled = "handleDfpDisabled"
         return (Data(), HTTPURLResponse())
     }
 
     // swiftlint:disable:next function_parameter_count
-    func handleDFPObservationMode(session _: URLSession, request _: URLRequest, publicToken _: String, captcha _: CAPTCHA, dfp _: DFPClient, requestHandler _: (URLSession, URLRequest) async throws -> (Data, HTTPURLResponse)) async throws -> (Data, HTTPURLResponse) {
+    func handleDFPObservationMode(session _: URLSession, request _: URLRequest, publicToken _: String, captcha _: CaptchaProvider, dfp _: DFPProvider, requestHandler _: (URLSession, URLRequest) async throws -> (Data, HTTPURLResponse)) async throws -> (Data, HTTPURLResponse) {
         methodCalled = "handleDFPObservationMode"
         return (Data(), HTTPURLResponse())
     }
 
     // swiftlint:disable:next function_parameter_count
-    func handleDFPDecisioningMode(session _: URLSession, request _: URLRequest, publicToken _: String, captcha _: CAPTCHA, dfp _: DFPClient, requestHandler _: (URLSession, URLRequest) async throws -> (Data, HTTPURLResponse)) async throws -> (Data, HTTPURLResponse) {
+    func handleDFPDecisioningMode(session _: URLSession, request _: URLRequest, publicToken _: String, captcha _: CaptchaProvider, dfp _: DFPProvider, requestHandler _: (URLSession, URLRequest) async throws -> (Data, HTTPURLResponse)) async throws -> (Data, HTTPURLResponse) {
         methodCalled = "handleDFPDecisioningMode"
         return (Data(), HTTPURLResponse())
     }
